* release/1.1.0:
  bump version to 1.1.0, regenerate javadocs
  enable manual approval of circleCI test step.
  add rate limit configuration options
  try with lower request rate
  add Rate limiting to Harvest API requests
  make tests more robust to existing data changing
  estimate item quantity is also a double
  throw RateLimitedException if we would have to wait more than 5 minutes to retry
  invoice item quantity is actually a double

+++ src/test/java/ch/aaap/harvestclient/impl/invoice/InvoicesApiUpdateTest.java
@@ -88,13 +88,13 @@ class InvoicesApiUpdateTest {
     @Test
     void changeLineItem() {
 
-        long quantity = 20;
+        double quantity = 20;
         double unitPrice = 30;
         String firstName = "test First";
         Invoice creationInfo = ImmutableInvoice.builder()
                 .client(clientReference)
                 .addInvoiceItem(ImmutableInvoiceItem.builder()
-                        .quantity(10L)
+                        .quantity(10.)
                         .unitPrice(1.)
                         .kind(kind)
                         .build())
@@ -109,7 +109,7 @@ class InvoicesApiUpdateTest {
 
         InvoiceUpdateInfo changes = ImmutableInvoiceUpdateInfo.builder()
                 .addInvoiceItem(ImmutableInvoiceItem.builder()
-                        .quantity(20L)
+                        .quantity(20.)
                         .unitPrice(2.)
                         .kind(kind)
                         .build())
